Match GlobalGameObject::RequestAction, GlobalGameObject::Receive, and GameStatus::IncLevel
- Fully implement GlobalGameObject::RequestAction (0x00416e20) and GlobalGameObject::Receive (0x00416e90) matching 100.00% (thunk), completing all 15 functions in GlobalGameObject.cpp.
- Adjust GameStatus::IncLevel (0x00408dc0) local register allocation to match 100.00% (thunk).
- Correct BaseCommonSocket::m_port to signed short based on MSVC mangling and 68K signatures, improving BaseCommonSocket::BaseCommonSocket (0x0045f680) to 85.71%.

+++ src/Visos/Network/WriteSocket.h
@@ -21,7 +21,7 @@ public:
 	// FUNCTION: LEMBALL 0x00462970
 	virtual void SetPort(short p_arg0) // vtable+0x28
 	{
-		m_port = (unsigned short) p_arg0;
+		m_port = p_arg0;
 	}
 	bool Send(NetworkMessage& p_arg0);
 	bool SendCritical(NetworkMessage& p_arg0);
